Updated to latest Rust: 'as uint' for array indexing

+++ visualizer/obj.rs
@@ -96,7 +96,7 @@ impl Model {
         let mut mesh = Vec::new();
         for face in self.faces.iter() {
             for i in range(0, 3) {
-                let vertex_id = face.vertex[i] - 1;
+                let vertex_id = face.vertex[i as uint] - 1;
                 mesh.push(*self.coords.get(vertex_id as uint));
             }
         }
@@ -107,7 +107,7 @@ impl Model {
         let mut tex_coords = Vec::new();
         for face in self.faces.iter() {
             for i in range(0, 3) {
-                let texture_coord_id = face.texture[i] as uint - 1;
+                let texture_coord_id = face.texture[i as uint] as uint - 1;
                 tex_coords.push(*self.texture_coords.get(texture_coord_id));
             }
         }
